Add retry handler for throttling and break tests into separate files by area
