+++ VVRestApi/VVRestApi/VVRestApi.csproj
@@ -3,14 +3,14 @@
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
     <PackageId>VVRestApi</PackageId>
-    <Version>5.1.2</Version>
-    <AssemblyVersion>5.1.2.0</AssemblyVersion>
+    <Version>5.1.3</Version>
+    <AssemblyVersion>5.1.3.0</AssemblyVersion>
     <Product>VisualVault REST API for .NET</Product>
     <Copyright>Copyright ©  2022</Copyright>
     <Description>VisualVault REST API for .NET</Description>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <Company>VisualVault</Company>
-    <PackageReleaseNotes>Added overload to RunWebService method that can take a usId to specify what user's preferences to use when executing the web service</PackageReleaseNotes>
+    <PackageReleaseNotes>Add automatic retry logic when receiving throttling response from VisualVault</PackageReleaseNotes>
     <RepositoryUrl>https://github.com/VisualVault/VisualVault.Net.RestClient</RepositoryUrl>
     <RepositoryType>GitHub</RepositoryType>
     <PackageLicenseExpression></PackageLicenseExpression>
